requirements.txt
Package             Version
------------------- ------------
certifi             2024.2.2
charset-normalizer  3.3.2
contourpy           1.2.1
cycler              0.12.1
filelock            3.13.4
fonttools           4.51.0
fsspec              2024.3.1
huggingface-hub     0.22.2
idna                3.7
importlib_resources 6.4.0
joblib              1.4.0
kiwisolver          1.4.5
matplotlib          3.8.4
numpy               1.26.4
packaging           24.0
pillow              10.3.0
pip                 23.3.1
pyparsing           3.1.2
python-dateutil     2.9.0.post0
PyYAML              6.0.1
requests            2.31.0
scikit-learn        1.4.2
scipy               1.13.0
setuptools          68.2.2
six                 1.16.0
threadpoolctl       3.4.0
timm                0.6.13
torch               1.10.0+cu111
torchvision         0.11.0+cu111
tqdm                4.66.2
typing_extensions   4.11.0
urllib3             2.2.1
wheel               0.41.2
zipp                3.18.1
